Move tags comps into folder
